Refactor germanized plugin asset loaded
